exit-code registry + oracle-routing recogniser: fix the scanners, not the tables
exit_codes (crates/no-shell-gate/tests/exit_codes.rs): 8 passed 5 failed -> 13 passed
0 failed on the rch lane (worktree, not a sha).

ROOT CAUSE, one line: the emission scanner masked comments and KEPT string literals. It
called `text_structure::code_only`, whose own doc says to use it "when string literals are
themselves the evidence being measured" — the `oracle_routing` case, and the exact opposite
of this one. So the half of the prose-is-not-an-emission defect that lives in a STRING
survived the fix that named it. Switched to `text_structure::code_and_literals` and applied
it to the `fn exit_code` body scan too, which is brace-matched across lines and had been
reading raw text. Four legs turned on that one choice:

  - a_code_that_appears_only_in_prose_is_not_an_emission: the fixture's own
    `let note = "we used to ...ExitCode::from(213)..."` was scored as an emission.
  - every_emitted_exit_code_is_documented: it was demanding rows for 138, 142, 306 and 307,
    read out of a comment in `ompo-doctor/src/health_repair.rs`'s `fn exit_code` body that
    cites `adapter_exec.rs:138-142` and `:306-307`. 306/307 are not representable in the
    `u8` that body returns. Documenting them would have been documenting fiction.
  - every_narrowing_exit_cast_has_a_reasoned_allowance: `cargo-lane-budget/src/lib.rs:989`
    quotes `ExitCode::from(selftest() as u8)` inside an assertion message.
  - every_pass_through_site_is_declared: same string manufactured a `selftest()` site.

The 14 REAL undocumented codes then got rows (5-11, 20-22, 30-32, 101), each naming its
emitters at file:line, what it does NOT mean, and the operator's action. 101 is recorded as
a standing violation of the range table rather than renumbered: `head_compiles` emits it for
a failed or timed-out HEAD build, its own unit test pins it at :405, and it is
indistinguishable from `XC-EXT-101`, a real panic. Ten undeclared pass-through expressions
got §6 rows; the header read "25 sites across 9 expressions" over a table of 10 and the
derived figure is 61 across 20.

Two subject fixes rather than allowance rows:
  - grader_attribution_gate::ledger_gate_exit returned i32 over a `{0,1}` range, so the
    caller wrote `ExitCode::from(ledger_exit as u8)`. Returns u8; the cast is gone. Removing
    a narrowing cast beats justifying one.
  - fleet-composite's `EXIT_USAGE = 64` collided with finding's `EXIT_USAGE = 2`
    (`crates/finding/src/main.rs:44`). Renamed to `EXIT_EX_USAGE`, NOT deleted: two distinct
    causes must stay distinguishable to a caller. The earlier `EXIT_CONFIG -> EXIT_USAGE`
    rename had moved this collision instead of ending it.

Three allowance rows were re-anchored, not relaxed: `:490 -> :492`, `:18 -> :282`,
`:853 -> :1146`. Each file holds exactly one occurrence of its expression; the sites moved
and the reviewed fact did not. The bidirectional check caught both halves (UNDECLARED at
the new line, ORPHANED at the old), which is the shape it was written for.

oracle_routing: 5 passed 4 failed -> 7 passed 2 failed at 15:39 on the lane, with the two
remaining reds now separated into one real defect and one environment blocker.

ROOT CAUSE: the two surfaces were centralised into `tick_monitor::TMUX` / `NTM`
(`crates/tick-monitor/src/kernel.rs:7,10`) and `surface_evidence` keyed only on the `"tmux"`
argv literal, so it went blind to every crate that adopted the shared constant.
`pane-oracle-diff` scored tmux=0 ntm=0 and `oracle-pane-state-differential` tmux=1 ntm=0 —
the two crates the known-good leg names as THE reference implementations of routing. Five
allowance rows read as STALE simultaneously and the message said to delete them; deleting
them would have removed five live DEBT exemptions for crates that do read both surfaces.
Added the shared-constant form, read from `code_and_literals` (a bare `NTM` inside an
assertion message is prose) while the three argv-shaped forms keep reading `code_only`
(an argv literal IS the evidence). The mask differs by form because the evidence does.

Four crates the fixed recogniser surfaced carry new reasoned rows, each with a dies-when:
tick-monitor (it DECLARES both names and spawns only tmux), pane-truth (a true reader — the
gate's header had cited it as a false positive, which was true of the substring instrument
and not of the crate), ompo-doctor (one liveness probe per binary, never two readings of one
fact), receiver-receipt (its `NTM` hit is a label inside an in-file `#[cfg(test)]` module).

Second defect, found by strengthening a message: `crate_sources` concatenated files in
`read_dir` order and callers masked the JOIN, so an unterminated `/*` in one file blanks
every file joined after it — `omp-orchestrator/src/input_closure.rs` holds 4 `/*` against 1
`*/` and `src/lib.rs` 5 against 3, legitimately, inside literals. Because `read_dir` order is
undefined, `omp-orchestrator` measured tmux=15 ntm=6 on APFS and tmux=0 ntm=0 on the worker
over a byte-identical 786,891-byte scan set. Now masked PER FILE in sorted order, which is
what `exit_codes::scan` already did and is immune by construction.

Messages strengthened where a message had sent a reader toward a weakening fix:
  - the stale-allowance leg now names three causes and MEASURES each row (roster membership,
    files, bytes, per-surface counts) instead of printing a bare name list.
  - the roster leg compares SETS and prints the symmetric difference instead of two integers.
    Strictly stronger: equal counts with different members used to pass.

text-structure gains one function, `identifier_occurrences`, and `has_identifier` now
delegates to it, so the identifier-boundary rule has one implementation and two callers
rather than a second copy in a gate test.

STILL RED, both named, neither forced:
  - the_crate_roster_is_derived_from_git_and_covers_the_tree — BLOCKED, not fixable by me.
    `crates/kernel-only-gate` and `crates/omp-host-tool-guard` are complete, untracked crate
    directories authored 2026-09-08/09 and workspace members through the `crates/*` glob, so
    they build and are invisible to every git-derived census including CI's. They are another
    agent's files; Main is routing them.
  - the lane cannot measure any git-derived roster reliably. Same leg, same commit, four
    answers: 92 tracked on this Mac, 85 at 15:04, 86 at 15:39, and 0 from 15:42 onward,
    where `git ls-files` fails outright and the anti-vacuity guards in all four legs fire
    correctly. A flapping count with no source diff is an instrument symptom. CI, which has a
    real `.git`, is the oracle for these four legs; the lane is not.

+++ crates/grader-attribution-gate/src/lib.rs
@@ -567,7 +567,13 @@ impl fmt::Display for CeilingVerdict {
 
 /// Exit 1 only on a ceiling breach. Slack and exact stay green so a clean
 /// ledger does not fail CI. Empty scan is 2, from the parser, not here.
-pub fn ledger_gate_exit(unattributed: &[String]) -> i32 {
+///
+/// Returns `u8`, the width `ExitCode::from` actually takes. It returned `i32` and the caller
+/// wrote `ExitCode::from(ledger_exit as u8)`, which is an exit-path narrowing cast over a
+/// value whose range is `{0,1}` — a cast that can never be exercised and still has to be
+/// justified to `exit_codes::every_narrowing_exit_cast_has_a_reasoned_allowance`. Removing
+/// the cast is cheaper and stronger than a reasoned allowance row for it.
+pub fn ledger_gate_exit(unattributed: &[String]) -> u8 {
     if CeilingVerdict::from_counts(unattributed.len(), UNATTRIBUTED_CLOSE_CEILING).refuses() {
         1
     } else {
